+++ CHANGELOG.md
@@ -203,13 +203,20 @@
 #####fixed
 1. Fixed can settings overlapping address not properly detected
 
+###v1.5.5
+#####added
+1.Hardware version under info.
+
+#####fixed
+1. AOUT calibration read and write.
+2. tool tip:Sensor not connected for TC.
+
 ###v1.6.0
 #####changed
 1. Electron@6.0.1
 
 #####added
-1.Hardware version under info.
 2.Support for multi USB connections.
 
 #####fixed
-1. AOUT calibration read and write.
+1. Calibration Import from json not working.